chore: move dataset buttons into previous existing component
